ref(sdk-crashes): Rename cocoa crash detector file
Rename cocoa_sdk_crash_detector to sdk_crash_detector_configs
to prepare having all SDK configs in one file.
